Fix all linter warnings

+++ src/org/edumips64/core/is/ALU_IType.java
@@ -107,7 +107,7 @@ public class ALU_IType extends ComputationalInstructions {
         int imm=-223;
         params.add(rt);  //destinazione R2
         params.add(rs);  //sorgente1 R3
-        params.add((int)imm);  //immediato R4
+        params.add(imm);  //immediato R4
         try {
            cpu.getRegister(rs).writeDoubleWord(337236854775807L); //rs register
            ins.setParams(params);
